safari-534.53-branch/LayoutTests/media/unsupported-rtsp.html


<video controls></video>

<p>Test that QuickTime file with RTSP URL generates a load error.<p>

<script src=video-test.js></script>
<script>
    video.src = "rtsp://a2047.v1411b.c1411.g.vq.akamaistream.net/5/2047/1411/2_h264_650/1a1a1ae454c430950065de4cbb2f94c226950c7ae655b61a48a91475e243acda3dac194879adde0f/wwdc_2006_2_650.mov";
    waitForEvent("error", function () {
        testExpected("video.error", null, "!=");
        testExpected("video.error.code", MediaError.MEDIA_ERR_SRC_NOT_SUPPORTED);
        testExpected("video.networkState", HTMLMediaElement.NETWORK_NO_SOURCE);
        endTest();
    });
</script>
